requirements.txt
pocoui
pytest
pytest-rerunfailures
pytest-xdist
allure-pytest
python-dotenv
logzero